+++ apps/etterlatte-saksbehandling-ui/client/src/App.tsx
@@ -24,7 +24,7 @@ import { mapResult } from '~shared/api/apiUtils'
 import { setDefaultOptions } from 'date-fns'
 import GenerellOppgave from '~components/generelloppgave/GenerellOppgave'
 import { VurderAktivitetspliktOppgave } from '~components/aktivitetsplikt/VurderAktivitetspliktOppgave'
-import { initAmplitude } from '~utils/amplitude'
+import { initAmplitude } from '~utils/analytics'
 import { Unleashcontext, useUnleash } from '~useUnleash'
 import { EtteroppgjoerForbehandling } from '~components/etteroppgjoer/forbehandling/EtteroppgjoerForbehandling'
 import { MeldtInnEndring } from '~components/meldtInnEndring/MeldtInnEndring'
