+++ apps/etterlatte-saksbehandling-ui/client/src/components/behandling/vilkaarsvurdering/KopierVilkaarAvdoed.tsx
@@ -4,7 +4,7 @@ import { mapResult } from '~shared/api/apiUtils'
 import { Alert, BodyShort, Box, Button, Heading, HStack, Link, Spacer, VStack } from '@navikt/ds-react'
 import { ExternalLinkIcon, PlusCircleIcon, XMarkOctagonIcon } from '@navikt/aksel-icons'
 import { ApiErrorAlert } from '~ErrorBoundary'
-import { ClickEvent, trackClickJaNei } from '~utils/amplitude'
+import { ClickEvent, trackClickJaNei } from '~utils/analytics'
 import { JaNei } from '~shared/types/ISvar'
 import {
   hentKandidatForKopieringAvVilkaar,
