+++ apps/etterlatte-saksbehandling-ui/client/src/components/oppgavebenk/oppgaveModal/oppfoelgingsOppgave/OpprettOppfoelgingsoppgaveModal.tsx
@@ -5,7 +5,7 @@ import { opprettOppgave } from '~shared/api/oppgaver'
 import { useInnloggetSaksbehandler } from '~components/behandling/useInnloggetSaksbehandler'
 import { FeatureToggle, useFeaturetoggle } from '~useUnleash'
 import { useForm } from 'react-hook-form'
-import { ClickEvent, trackClick } from '~utils/amplitude'
+import { ClickEvent, trackClick } from '~utils/analytics'
 import { OppgaveKilde, Oppgavetype } from '~shared/types/oppgave'
 import { BodyShort, Box, Button, Modal, Textarea, VStack } from '@navikt/ds-react'
 import { isPending } from '~shared/api/apiUtils'
