+++ apps/etterlatte-saksbehandling-ui/client/src/components/person/sakOgBehandling/EndreEnhet.tsx
@@ -8,7 +8,7 @@ import { PencilIcon } from '@navikt/aksel-icons'
 import { ApiErrorAlert } from '~ErrorBoundary'
 import { useInnloggetSaksbehandler } from '~components/behandling/useInnloggetSaksbehandler'
 import { useForm } from 'react-hook-form'
-import { ClickEvent, trackClick } from '~utils/amplitude'
+import { ClickEvent, trackClick } from '~utils/analytics'
 
 interface EndreEnhetSkjema {
   enhet: string
