+++ apps/etterlatte-saksbehandling-ui/client/src/shared/statusbar/Statusbar.tsx
@@ -15,7 +15,7 @@ import { Navbar } from '~shared/header/Navbar'
 import { useAppDispatch } from '~store/Store'
 import { settPerson } from '~store/reducers/PersonReducer'
 import { PersonOversiktFane } from '~components/person/Person'
-import { ClickEvent } from '~utils/amplitude'
+import { ClickEvent } from '~utils/analytics'
 import { ApiErrorAlert } from '~ErrorBoundary'
 import { hentSakMedBehandlnger } from '~shared/api/sak'
 import { AdressebeskyttelseGraderingTag } from '~shared/tags/AdressebeskyttelseGraderingTag'
