+++ apps/etterlatte-saksbehandling-ui/client/src/utils/analytics.ts
